style: rename classes and corresponding files
