+++ wrapper/bridge.py
@@ -3,9 +3,9 @@ import argparse
 
 import tools.messages as msg
 
-from models.bundle import BundleCreator
-from models.kernel import KernelBuilder
-from models.assets import AssetCollector
+from models.bundle_creator import BundleCreator
+from models.kernel_builder import KernelBuilder
+from models.assets_collector import AssetsCollector
 
 from utils import Resources
 
@@ -93,7 +93,7 @@ def main(args: argparse.Namespace) -> None:
                 ksu = args.ksu,
             ).run()
         case "assets":
-            AssetCollector(
+            AssetsCollector(
                 codename = args.codename,
                 base = args.base,
                 chroot = args.chroot,
