+++ wrapper/models/__init__.py
@@ -1,3 +1,3 @@
-from .kernel import KernelBuilder
-from .assets import AssetCollector
-from .bundle import BundleCreator
+from .kernel_builder import KernelBuilder
+from .bundle_creator import BundleCreator
+from .assets_collector import AssetsCollector
